Add support for Oracle Linux instances and OCI CCM
This PR refactors some of the resources to create a k3s_primary_server
which is then used to optionally deploy the OCI Cloud Controller Manager
to the cluster. This is disabled by default.

A terraform.tfvars.sample is provided in the example/ folder to explain
some of the new/changed/removed variables.

Signed-off-by: Avi Miller <avi.miller@oracle.com>